small changes according to the review

+++ backend/app/comment_handler.go
@@ -114,6 +114,10 @@ func (app *App) deleteComment(c *gin.Context) (interface{}, Response) {
 		return nil, InternalServerError(errInternalServerError)
 	}
 
+	if fmt.Sprint(userID) != fmt.Sprint(comment.UserID) {
+		return nil, Forbidden(errors.New("you have no access to delete the comment"))
+	}
+
 	err = app.DB.DeleteComment(uint(id))
 
 	if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -127,9 +131,6 @@ func (app *App) deleteComment(c *gin.Context) (interface{}, Response) {
 		return nil, InternalServerError(errInternalServerError)
 	}
 
-	if fmt.Sprint(userID) != fmt.Sprint(comment.UserID) {
-		return nil, Forbidden(errors.New("you have no access to delete the comment"))
-	}
 	return ResponseMsg{
 		Message: "comment is deleted successfully",
 	}, Ok()
@@ -140,7 +141,13 @@ func (app *App) listComments(c *gin.Context) (interface{}, Response) {
 	bugIDStr := c.Query("bug_id")
 	UserID := c.Query("user_id")
 
-	bugID, _ := strconv.ParseUint(bugIDStr, 10, 64)
+	var bugID uint64
+
+	if bugIDStr != "" {
+
+		bugID, _ = strconv.ParseUint(bugIDStr, 10, 64)
+	}
+
 	comments := app.DB.ListComments(uint(bugID), UserID)
 
 	if len(comments) == 0 {
@@ -167,6 +174,10 @@ func (app *App) updateComment(c *gin.Context) (interface{}, Response) {
 	}
 
 	idStr := c.Param("id")
+
+	if idStr == "" {
+		return nil, BadRequest(errors.New("comment id is required"))
+	}
 
 	id, _ := strconv.ParseUint(idStr, 10, 64)
